+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 
 - Correctly enqueue deferred tasks with only positional arguments on Ruby < 3.3.4 (#385).
+- Improve restart logs in daemon (#391).
 
 ## [1.27.0] - 2026-08-06